implemented Event endpoint. NOTE: in DB user.username is now UNIQUE

+++ spring-server/src/main/java/ch/heigvd/gamification/database/dao/PointScaleRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 /**
  * Created by lux on 10.12.16.
  */
-public interface PointScaleRepository extends CrudRepository<PointScale, Integer> {
+public interface PointScaleRepository extends CrudRepository<PointScale, Long> {
     List<PointScale> findByApplication(Application application);
     PointScale findByIdAndApplication(long id, Application application);
 }
